fix: fix auth bug user can get all accounts

+++ src/user/user.controller.ts
@@ -10,7 +10,10 @@ import {
   ValidationPipe,
   HttpCode,
   UseGuards,
+  Req,
+  UnauthorizedException,
 } from '@nestjs/common';
+import { Request } from 'express';
 import { UserService } from './user.service';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { AccessTokenGuard } from 'src/common/guards/access-token.guard';
@@ -21,7 +24,10 @@ export class UserController {
 
   @UseGuards(AccessTokenGuard)
   @Get(':id')
-  public async findOne(@Param('id') id: string) {
+  public async findOne(@Param('id') id: string, @Req() req: Request) {
+    if (id != req.user['sub']) {
+      throw new UnauthorizedException('Administrator privileges required');
+    }
     const user = await this.userService.findOne(+id);
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
     const { refreshToken, ...safeUser } = user;
@@ -32,9 +38,13 @@ export class UserController {
   @UseGuards(AccessTokenGuard)
   @Patch(':id')
   public async update(
+    @Req() req: Request,
     @Param('id') id: string,
     @Body() updateUserDto: UpdateUserDto,
   ) {
+    if (id != req.user['sub']) {
+      throw new UnauthorizedException('Administrator privileges required');
+    }
     const isUser = await this.userService.findOne(+id);
     if (!isUser) {
       throw new BadRequestException('User not found');
@@ -51,7 +61,10 @@ export class UserController {
   @UseGuards(AccessTokenGuard)
   @Delete(':id')
   @HttpCode(204)
-  public async remove(@Param('id') id: string) {
+  public async remove(@Param('id') id: string, @Req() req: Request) {
+    if (id != req.user['sub']) {
+      throw new UnauthorizedException('Administrator privileges required');
+    }
     const updatedUser = await this.userService.remove(+id);
     if (!updatedUser) {
       throw new BadRequestException("User didn't delete");
